backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.15.2
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
redis==5.2.1
arq==0.26.3
httpx==0.28.1
openai==1.75.0
aiosmtplib==3.0.2
python-multipart==0.0.20
email-validator==2.2.0
orjson==3.10.16
tenacity==9.1.2

pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
ruff==0.11.5
asgi-lifespan==2.1.0
